Update delete command
- allow user to delete saved groups before loading them

+++ src/main/java/seedu/duke/GroupCommand.java
@@ -5,6 +5,7 @@ package seedu.duke;
 import seedu.duke.exceptions.GroupDeleteException;
 import seedu.duke.storage.FileIO;
 import seedu.duke.storage.FileIOImpl;
+import seedu.duke.storage.GroupNameChecker;
 import seedu.duke.storage.GroupStorage;
 
 import java.util.Optional;
@@ -36,31 +37,22 @@ public class GroupCommand {
      * @param groupName the name of the group to delete
      */
     public static void deleteGroup(String groupName) {
-        if (Group.groups.containsKey(groupName)) {
-            Group.groups.remove(groupName);
-            deleteGroupFile(groupName);
-            System.out.println("The group " + groupName + " has been deleted.");
-        } else {
-            System.out.println("The group " + groupName + " does not exist.");
-        }
-    }
-
-    //@@author hafizuddin-a
-    /**
-     * Deletes the group file for the specified group name.
-     *
-     * @param groupName the name of the group whose file needs to be deleted
-     */
-    private static void deleteGroupFile(String groupName) {
         try {
             FileIO fileIO = new FileIOImpl();
             GroupStorage groupStorage = new GroupStorage(fileIO);
-            groupStorage.deleteGroupFile(groupName);
+            GroupNameChecker groupNameChecker = new GroupNameChecker();
+            if (groupNameChecker.doesGroupNameExist(groupName)) {
+                groupStorage.deleteGroupFile(groupName);
+                Group.groups.remove(groupName);
+                System.out.println("The group " + groupName + " has been deleted.");
+            } else {
+                System.out.println("The group " + groupName + " does not exist.");
+            }
         } catch (GroupDeleteException e) {
-            System.out.println("Failed to delete the group file: " + e.getMessage());
+            System.out.println("Failed to delete the group: " + e.getMessage());
         }
     }
-
+    //@@author hafizuddin-a
     /**
      * Adds a member with the specified name to the current group.
      * If the user is not currently in a group, prints a message asking them to create or join a group first.
